+++ lib/bibliothecary.rb
@@ -9,13 +9,9 @@ require "bibliothecary/runner"
 require "bibliothecary/exceptions"
 require "bibliothecary/file_info"
 require "bibliothecary/related_files_info"
-require "bibliothecary/purl_util"
 require "find"
 require "tomlrb"
 
-Dir[File.expand_path("bibliothecary/multi_parsers/*.rb", __dir__)].each do |file|
-  require file
-end
 Dir[File.expand_path("bibliothecary/parsers/*.rb", __dir__)].each do |file|
   require file
 end