+++ lib/bibliothecary/parsers/homebrew.rb
@@ -20,8 +20,6 @@ module Bibliothecary
         }
       end
 
-      add_multi_parser(Bibliothecary::MultiParsers::CycloneDX)
-      add_multi_parser(Bibliothecary::MultiParsers::DependenciesCSV)
 
       def self.parse_brewfile(file_contents, options: {})
         source = options.fetch(:filename, 'Brewfile')