+++ lib/bibliothecary/parsers/pub.rb
@@ -20,8 +20,6 @@ module Bibliothecary
         }
       end
 
-      add_multi_parser(Bibliothecary::MultiParsers::CycloneDX)
-      add_multi_parser(Bibliothecary::MultiParsers::DependenciesCSV)
 
       def self.parse_yaml_manifest(file_contents, options: {})
         manifest = YAML.load file_contents