+++ spec/bibliothecary_spec.rb
@@ -17,6 +17,7 @@ describe Bibliothecary do
           Bibliothecary::Parsers::Clojars,
           Bibliothecary::Parsers::CocoaPods,
           Bibliothecary::Parsers::Cog,
+          Bibliothecary::Parsers::Conan,
           Bibliothecary::Parsers::Conda,
           Bibliothecary::Parsers::CPAN,
           Bibliothecary::Parsers::CRAN,
@@ -414,12 +415,12 @@ describe Bibliothecary do
 
   it "allows resetting of config options" do
     Bibliothecary.configure do |config|
-      config.carthage_parser_host = "http://foobar.com"
+      config.ignored_dirs = ["custom_dir"]
     end
 
     Bibliothecary.reset
 
-    expect(Bibliothecary.configuration.carthage_parser_host).to eq("https://carthage.libraries.io")
+    expect(Bibliothecary.configuration.ignored_dirs).to eq([".git", "node_modules", "bower_components", "vendor", "dist"])
   end
 
   it "properly ignores directories based on ignored_dirs" do